# api/swagger/swagger.json
{
  "info": {
    "title": "Amazing Online Profiles!",
    "version": "1.0.0",
    "description": "Technical Challenge for Holiday Extras"
  },
  "host": "localhost",
  "basePath": "/api",
  "swagger": "2.0",
  "paths": {
    "/allusers": {
      "get": {
        "description": "Gets all users in the system",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "responses": {
          "200": {
            "description": "all users returned"
          },
          "404": {
            "description": "not found error"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/user": {
      "get": {
        "description": "Allows a user to get their profile details",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "_id",
            "description": "The mongo database id for the user",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "users details returned"
          },
          "404": {
            "description": "not found error"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/delete": {
      "delete": {
        "description": "Allows a user to update their profile",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "_id",
            "description": "The mongo database id for the user",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "deleted account"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/update": {
      "patch": {
        "description": "Allows a user to update their profile",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "_id",
            "description": "The mongo database id for the user",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "description": "It was in the requirements, but mongo uses _id. It'll be a special number.",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "email",
            "description": "Email address to use for login.",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "givenName",
            "description": "User's given name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "familyName",
            "description": "User's family name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "User's password.",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "about",
            "description": "Information about a user",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "updated account"
          },
          "400": {
            "description": "details failed validation"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/register": {
      "post": {
        "description": "Allows a user to register for an awesome profile",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "It was in the requirements, but mongo uses _id. It'll be a special number.",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "email",
            "description": "Email address to use for login.",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "givenname",
            "description": "User's given name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "familyname",
            "description": "User's family name",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "User's password.",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "about",
            "description": "Information about a user",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "created account"
          },
          "400": {
            "description": "details failed validation"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "Login to the application",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "email",
            "description": "Email address to use for login.",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "description": "User's password.",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "login"
          },
          "400": {
            "description": "login details failed validation"
          },
          "401": {
            "description": "login details failed authentication"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
    "/api-docs": {
      "get": {
        "description": "Displays api docs",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "responses": {
          "200": {
            "description": "Display api docs"
          }
        }
      }
    }
  },
  "definitions": {},
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}